+++ dask/dataframe/io/tests/test_parquet.py
@@ -13,7 +13,7 @@ import dask
 import dask.multiprocessing
 import dask.dataframe as dd
 from dask.dataframe.utils import assert_eq
-from dask.dataframe.io.parquet import _parse_pandas_metadata
+from dask.dataframe.io.parquet.utils import _parse_pandas_metadata
 from dask.utils import natural_sort_key
 
 try:
