+++ dask/dataframe/optimize.py
@@ -26,7 +26,7 @@ def optimize(dsk, keys, **kwargs):
         dsk, dependencies = cull(dsk, [keys])
     dsk = fuse_getitem(dsk, dataframe_from_ctable, 3)
     if fastparquet:
-        from .io.parquet import _read_parquet_row_group
+        from .io.parquet.fastparquet import _read_parquet_row_group
         dsk = fuse_getitem(dsk, _read_parquet_row_group, 4)
 
     dsk, dependencies = fuse(dsk, keys, dependencies=dependencies,
